Clarify the Commentary section
Combine the version-split Description subsection to one. Assume that the
latest version is now used.

Rename the Changes subsection to Notable Changes. The purpose of this
subsection is to describe shortly what has changed in each major and minor
version bumps. It should not compete with the git-log with detail.