Use template approach to grep's arguments
Previously the search pattern was just appended at the end of the command
line. Now the command line can contain a string ":search-pattern" and the
input from helm is replaces that string in the real command.

Previously the undocumented :extra-arguments property was used to convey the
search pattern. Now the :search-pattern is used for that. The behaviour of
:extra-arguments doesn't change. I.e. that string is appended to the command
line (if defined).

This fixes the issue where the search pattern could not start with a dash.