Fix escaping double-quotes when constructing the grep command
This caused the following error when double-quotes were given in the
helm-pattern:

(error "Candidates function ‘helm-grepint-grep-process’ should run a process")

When debugging is enabled, the underlying error is:

"Invalid use of `\\` in replacement text"

in (replace-match). I.e. regular expression replacer doesn't accept the
replacement string. The solution is to dumb it down to a simple replace (as
nothing fancier is needed in this case).

Fixes #16.